Refactor marketing homepage and resolve build blockers
