+++ apps/web/src/components/editor/LayersPanel.tsx
@@ -41,7 +41,7 @@ export function LayersPanel() {
       </header>
       <div className="space-y-3">
         {layers.map((layer, index) => {
-          const textLayer = isTextLayer(layer);
+          const textLayer = isTextLayer(layer) ? layer : null;
           const isActive = layer.id === activeLayerId;
 
           return (
